# Optimize Markdown formatter regex matching